refactor: move collapse btn from purchases
