+++ src/component/CollapseBtn.tsx
@@ -1,15 +1,16 @@
-import { IconButton } from "@chakra-ui/react"
+import { IconButton, IconButtonProps } from "@chakra-ui/react"
 import { Dispatch, FC, SetStateAction } from "react"
 import { FiChevronDown, FiChevronUp } from "react-icons/fi"
 
 interface CollapseBtnProps {
+  size?: IconButtonProps["size"]
   isExpanded: boolean
   setIsExpanded: Dispatch<SetStateAction<boolean>>
   isDisabled?: boolean
 }
 
 export const CollapseBtn: FC<CollapseBtnProps> = (props) => {
-  const { isExpanded, setIsExpanded, isDisabled } = props
+  const { size = "sm", isExpanded, setIsExpanded, isDisabled } = props
 
   const handleIsExpandedChange = () => {
     setIsExpanded((prevIsExpanded) => !prevIsExpanded)
@@ -20,7 +21,7 @@ export const CollapseBtn: FC<CollapseBtnProps> = (props) => {
   return (
     <IconButton
       aria-label="expand-row-btn"
-      size="sm"
+      size={size}
       variant="ghost"
       icon={btnIcon}
       onClick={handleIsExpandedChange}
