+++ src/component/purchaseHistory/PurchaseHistoryTableRow.tsx
@@ -3,7 +3,7 @@ import { PurchaseStatusBadge } from "component/badge/PurchaseStatusBadge"
 import { PurchaseDocumentsModal } from "component/document/PurchaseDocumentsModal"
 import { PurchaseGoodsModal } from "component/purchase/PurchaseGoodsModal"
 import { PurchaseSupplierModal } from "component/purchase/PurchaseSupplierModal"
-import { CollapseBtn } from "component/purchaseHistory/CollapseBtn"
+import { CollapseBtn } from "component/CollapseBtn"
 import { DeliveriesHistoryTable } from "component/purchaseHistory/DeliveryHistoryTable"
 import { PurchaseHistoryRowMenu } from "component/purchaseHistory/PurchaseHistoryRowMenu"
 import { PurchaseServicesModal } from "component/purchaseService/PurchaseServicesModal"
